sift: remove old sift program and replace with vlfeat
The previous sift implementation would fail with the following error:

```
sift: util.c:330: ConvHorizontal: Assertion `cols + kzsize < 8000`
failed.
```

As such, we're going to be using a different sift implementation than
before. The implementation that we've decided on is called "vlfeat", and
is available under the BSD License. Source code for the implementation
can be found at https://github.com/vlfeat/vlfeat.